+++ RELEASENOTES.md
@@ -48,6 +48,8 @@
         ([#9346](https://github.com/google/ExoPlayer/issues/9346)).
     *   Fix RTSP Session header handling
         ([#9416](https://github.com/google/ExoPlayer/issues/9416)).
+    *   Fix RTSP WWW-Authenticate header parsing
+        ([#9428](https://github.com/google/ExoPlayer/issues/9428)).
 *   UI:
     *   Use `defStyleAttr` when obtaining styled attributes in
         `StyledPlayerView`, `PlayerView` and `PlayerControlView`